src/demo1/requirements.txt
# =============================================================================
# Requirements for Qwen2 SFT Training with LoRA
# =============================================================================
# Install: pip install -r requirements.txt

# -----------------------------------------------------------------------------
# Core Dependencies
# -----------------------------------------------------------------------------
torch>=2.1.0
transformers>=4.40.0
datasets>=2.14.0
peft>=0.10.0
trl>=0.8.0
accelerate>=0.27.0

# YAML config support
pyyaml>=6.0

# -----------------------------------------------------------------------------
# Quantization (4-bit/8-bit training)
# -----------------------------------------------------------------------------
bitsandbytes>=0.41.0

# -----------------------------------------------------------------------------
# Flash Attention (faster training, requires CUDA)
# -----------------------------------------------------------------------------
# Uncomment if you have CUDA and want faster attention
# Note: Requires compilation, may need: pip install flash-attn --no-build-isolation
# flash-attn>=2.5.0

# -----------------------------------------------------------------------------
# Experiment Tracking
# -----------------------------------------------------------------------------
tensorboard>=2.15.0

# Optional: Weights & Biases
# wandb>=0.16.0

# Optional: MLflow
# mlflow>=2.10.0

# -----------------------------------------------------------------------------
# Development (optional)
# -----------------------------------------------------------------------------
# ipython>=8.0.0
# jupyter>=1.0.0

